Expose proof compression/decompression in wasm bindings
