+++ crates/js/whatever.mjs
@@ -1,0 +1,17 @@
+import { compress_proof } from "./pkg/semaphore_rs_js.js"
+
+const proof = [
+  "2D77679B613036865F4518894C80691CF65338FE7834FE3DD5F98C4F0F5A9E6D",
+  "24018E845EDF74D69528A63EED053296A397DF13A1D08873E2B2D673837B31C3",
+  "99D39B2CBCA524B5916AC97DBC4AFC1B8A5F59D65BA583FC49EC2677226E926",
+  "DA5812D7B4E0BEB22D25C194431674396AEC70751873EDB9AC8C933BA1F0F2E",
+  "723CACA23EFB9AA44DB59EAD0EEB28C2EFB9C766D9A3F994ED047179E37B347",
+  "2166D9FC2D4CF446B120E5663880E0927825AA36A02B896AC0F3A5EF6E0239B",
+  "287FB1D0415A734BA76DF9EB50CA6758BB806272F8FE40E3ADBAD3A850C05167",
+  "1240CF8AA43CF4EA4A2D8DFFAC653A6467CEFD0F19E129CFFAD85299D6705444",
+]
+
+const compressed = compress_proof(proof)
+
+console.log(compressed)
+
