+++ crates/proof/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "semaphore-rs-proof"
+version.workspace = true
+edition.workspace = true
+homepage.workspace = true
+license.workspace = true
+repository.workspace = true
+authors.workspace = true
+description.workspace = true
+keywords.workspace = true
+categories.workspace = true
+
+[dependencies]
+semaphore-rs-utils.workspace = true
+
+ruint.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+ark-circom = { workspace = true, optional = true }
+ark-ec = { workspace = true, optional = true }
+ark-groth16 = { workspace = true, optional = true }
+ark-bn254 = { workspace = true, optional = true }
+lazy_static.workspace = true
+getrandom.workspace = true
+alloy-core.workspace = true
+hex.workspace = true
+
+[features]
+default = ["ark"]
+ark = ["dep:ark-circom", "dep:ark-ec", "dep:ark-groth16", "dep:ark-bn254"]
+
