+++ crates/proof/src/compression.rs
