+++ crates/utils/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "semaphore-rs-utils"
+version.workspace = true
+edition.workspace = true
+homepage.workspace = true
+license.workspace = true
+repository.workspace = true
+authors.workspace = true
+description.workspace = true
+keywords.workspace = true
+categories.workspace = true
+
+[dependencies]
+serde.workspace = true
+tiny-keccak.workspace = true
+hex.workspace = true
+
+[dev-dependencies]
+serde_json.workspace = true
+bincode.workspace = true
+
